refactor(auth): Split hashers into per-algorithm submodules
- Sort PasswordHashingAlgo variants alphabetically
- Move each hasher into its own file under password_hashing/
  with colocated unit tests
- Rewrite stale BcryptSha256 salt comment; explain the
  canonical Radix64 encoding invariant clearly
- Centralize salt generation into generate_salt() in mod.rs
- Replace all em-dash characters with ASCII hyphens
- Restore Cargo.lock to upstream/password_hash baseline
  (eliminates openraft alpha.23 version skew)

Note: This commit was done with the help of AI
