+++ crates/config/src/identity.rs
@@ -70,10 +70,10 @@ pub enum PasswordHashingAlgo {
     Bcrypt,
     /// Bcrypt combined with SHA256.
     BcryptSha256,
-    /// Scrypt.
-    Scrypt,
     /// PBKDF2 with SHA512.
     Pbkdf2Sha512,
+    /// Scrypt.
+    Scrypt,
     // #[cfg(test)]
     /// None. Should not be used outside of testing where expected value is
     /// necessary.
